refactor(mm-next): move componet `wine-warning` file location
